Fix AppearanceGenerator to work for files with invalid /AP dicts for check boxes
Some applications (tested with Okular) save a PDF that contains a check
box and where the appearances should be regenerated in a manner that -
in my understanding - contradicts the specification.

A check box's /AP should have a subdictionary containing an /Off key and
another key like /Yes for the on state. However, a PDF saved with
Okuklar has the /AS set to /N (instead of /Off or /Yes) and an /AP value
of <</N ref>> where ref refers to an appearance stream.

This prohibits HexaPDF to correctly regenerate the appearances. To
remedy this, the /N key is now checked for button fields and corrected
in case it doesn't refer to a subdictionary.