Add configuration option encryption.on_decryption_error and relax error handling
HexaPDF already handles some decryption problems in a user-friendly way
by ignoring them if possible. However, the problem when not enough data
is given, currently always leads to an exception.

With this new configuration option the behaviour for the case of 'not
enough data' can now be customized. And the default handling has been
relaxed to just try and decrypt the data anyways (the reasoning behind
being that the result does not get worse).