Implement Dictionary#to_hash and update minimum required Ruby to 3.0
When a Hash object is compared to a Dictionary, the comparison always
fails because the Hash object doesn't know how to compare itself to a
Dictionary. It is no problem the other way around.

However, there is a work-around: By implementing #to_hash the
Hash#==(other) method delegates to other#== to do the comparison. This
is similar to how Array#== works in conjunction with #to_ary. Note that
other is only checked for the existence of the #to_hash method, that
method is never actually called.

The Dictionary class originally had a #to_hash method. However, this
method was removed in 2017 in favour of #to_h because #to_hash made
Dictionary objects destructurable when using them as method arguments.

Since handling of keyword arguments has changed in Ruby 3.0, an existing
Dictionary#to_hash method won't destructure the object anymore. And we
have the benefit that the comparison of Hash objects with Dictionary
objects works regardless of their order. This is important, for example,
when determining the objects for an incremental update.