Allow adding custom XMP metadata
The XMP support in HexaPDF is rather limited. Though it supports
registering custom namespaces and properties, not everything is
supported because the RDF format provides many ways to actually define
properties.

This change adds support for adding custom raw metadata to the generated
XMP document. The user has to ensure that the custom metadata is valid
and conforms to the specification. With this change it is possible to
add custom metadata like PDF/A schema definitions to the metadata.