Refactor the box implementation of the document layout system
The document layout system works quite well but still has some
deficiencies like:

* Box subclasses sometimes implement common code like checking whether
  the specified width/height fit into the available space

* Not all box subclasses rely on the common behaviour defined in Box

* Overflowing boxes need to be handled by subclasses

This change tries to remedy the situation somewhat so that common code
is in the base Box class and future features can more easily be
integrated.

Note that the changes mostly affect only the box implementations
themselves and not their usage which stays the same.

The biggest change is that the FitResult class moved from Frame to Box
and that each box instance now has an associated FitResult instance.
This allows the box instance to pre-fill the fit result and also allows
to communicate more data to Frame.

All subclasses now need to implement #fit_content instead of #fit and
the latter now actually returns the associated FitResult with the
appropriate result status.

The fit result status also changed from purely true/false to the actual
result and can currently be :success, :failure or :overflow. The latter
allows the layout engine to deal with overflowing boxes without the
boxes themselves doing more work.

By relying on the new fit result status, the Box#split method is now
also much clearer and easier to understand. And subclasses just need to
implement the #split_content method for the case of actually splitting a
box into two.

Additionally, during the refactoring some bugs related to the individual
implementations of common logic were revealed and fixed.