Update Layout::InlineBox to allow usage without predefined width
The InlineBox currently needs a width set to work correctly. However,
this is not strictly necessary, e.g. for wrapped text boxes.

This change makes it so that the width doesn't need to be set on
creation but only by the wrapped box on fitting. Furthermore,
\#fit_wrapped_box is simplified and doesn't use Frame#fit anymore. This
results in a correct mask box and simplified drawing code.