Fix Importer#import to correctly import streams backed by FiberDoubleForString
The FiberDoubleForString class is used when a string should be returned
by a filter but the string is not yet available or maybe need to be
adjusted before returning (e.g. Canvas#stream_data).

Since an instance of this class may return the same string every time,
we need to duplicate the underlying string instead of the stream data
object itself.