Fix Layout::PageStyle#create_page to always return new frames and update API docs
The #create_page method doesn't set a #frame object on each
invocation (if the template doesn't do it). This leads to the same Frame
being used for all pages created from the page style which doesn't work
because once the frame is full, nothing can be drawn anymore and an
endless loop is entered.

This change makes it so that #create_page always sets a new #frame
object on each invocation.