Fix Layout::TableBox row height calculation and distribution wrt to row spans
The current implementation of the table fitting algorithm does not
consider the distribution of row heights for row span cells at all. This
leads to bigger than necessary tables.

Consider this table which is rendered like the drawing on the right:

---------      ---------
| a | b |      | a | b |
| a |   |      | a |   |
| a |---|      | a |   |
| a | c |      | a |   |
| a |   |      | a |   |
---------      |   |---|
               |   | c |
               ---------

This change corrects the calculation and distribution of the row heights
to reduce the size of the height of the table and create a more expected
version of the table.