Added configuration option 'acro_form.fallback_default_appearance'
The PDF specification requires the /DA to be available. However, there
are invalid PDF files that have text fields without a /DA entry and also
no /DA entry on the main interactive form dictionary.

Prior to this change an error was raised in such a situation when the
information from the /DA was needed. But this might not always be the
best choice of action.

This change introduces a new configuration option that, by default,
makes HexaPDF add and use a standard default appearance for variable
text fields without a /DA entry. The old behaviour can be activated by
setting this variable to nil.