Change optimization task to not remove optional /Type entries containing default values
An optional /Type field is currently removed like any other optional
field if it contains its default value.

Since this can lead to problems this change makes the optimization task
leave the /Type entries alone. For example, some Factur-X validators
need /Type /EmbeddedFile entries for embedded files because if this
entry is missing they declare the PDF non-conforming.

It is possible to work around this by not optimizing the file or
optimizing it and then adding the necessary keys back. But keeping the
/Type entry also has other benefits like automatic type wrapping in the
general case when such a file is loaded by HexaPDF again.