Fix handling of null values when importing objects
If a PDF object is deleted from a document, it is usually only marked as
free. This means its value is set to `nil` and the object reference
removed the document store.

However, the null object might already be referenced by another object
and if the null object is a Dictionary, accessing many of its methods
will lead to an error, as is the case when importing.

This can be fixed by explicitly handling null values.