Remove /Type entry from trailer
When a document is optimized and xref streams are deleted, the
associated trailer might still have the /Type entry set. This leads to
the written trailer dictionary having the /Type entry which might lead
to problems (because the trailer is a standard trailer and not an xref
stream).

There are other ways in which a trailer might have a lingering /Type
entry.

This change deletes the /Type entry from the trailer.