Remove setting of /DA of AcroForm::Form validation
Setting a default appearance string on the main AcroForm dictionary
conflicts with the immutability expectation of digitally signing a
document (e.g. if a document is already signed and a second signature is
added with HexaPDF, the additional /DA entry is interpreted as a not
allowed change).

Since HexaPDF will now automatically set a default appearance string on
a field if it is needed (e.g. for creating a visual appearance), the
general setting of an apperance string in the main AcroForm dictionary
is also not necessary anymore.