Change text-like AcroForm field creation methods to always set /DA
Due to recent changes no default /DA is set on the main AcroForm
dictionary on validation anymore. This means that if no font information
is provided on field creation, neither a field specific /DA nor a
default /DA exists. However, the PDF spec requires a /DA to be set.

If a value is assigned to a field, a fallback mechanism automatically
sets a /DA but this case should only be relied on as a backup in case of
working with invalid PDFs.

Therefore a default appearance string is now always set when creating
text-like fields.