Add task for merging AcroForm information for imported pages
When pages from PDF are merged to another PDF, only the basic
information reachable from the page itself is retained, for example, the
annotations.

If a page contains AcroForm field widgets, the widgets are also retained
since they are annotations. However, the fields themselves are not
correctly merged because they primarily live elsewhere.

This change adds a task for merging the AcroForm information of imported
widgets/fields.