Small optimization for outputting TextFragments
When we are at the start of showing text in a TextFragment (i.e. the
`canvas.begin_text` actually starts a text object), the text line matrix
and the text matrix are the identity matrices. This means that we don't
need to use the SetTextMatrix operator but can use the relative MoveText
operator.