Fix parsing indirect objects with indirect references as values
According to the PDF specification:

    NOTE (2020) Any object outside of an object stream (see 7.5.7,
    "Object streams") can consist solely of an object reference. PDF
    syntax thus permits chains of such objects.

This means that indirect objects that are not object streams can look
like this:

    3 0 obj
    5 0 R
    endobj

This was confirmed in a comment on the pdf-issues repository, see
https://github.com/pdf-association/pdf-issues/issues/102#issuecomment-2395655111

This change allows HexaPDF to follow such chains of indirect
references. Note that HexaPDF still errors out if a reference is used
for a stream object.