Special handling of subset glyph IDs is not needed anymore
The special handling of the subset glyph IDs where 4 numbers were left
out (13, 40, 41, 92) is not needed anymore. This was originally done
because the subset glyph IDs were directly used as char codes and those
four numbers represent characters that would need to be escaped in
literal strings.

Due to a recent change the mapping of char codes to (subset) glyph IDs
is now done differently.