Fix PDF/A validation and Adobe Reader display problems due to missing required fields
The /Encoding entry in a Type0 font needs to be a Symbol or a valid CMap
stream. These streams needs additional entries that were missing since
the the recent change for supporting multiple codepoints referencing the
same glyph introduced them.

This lead to a PDF/A validation regression and a display problem on
Adobe Reader.