Fix font wrappers to handle the case of deleted objects
When a font wrapper is created, an associated PDF object is created and
added to the document. During processing of the document it may happen
that the added font is never used and therefore deleted. When during
writing the font's associated PDF object is then completed, it doesn't
work because it is already null.