Add argument 'compact' to Document#write
According to the PDF specification the initial revision of a document
has to contain a single cross-reference subsection that starts at
zero. All entries (except the one for zero) in that subsection are
assumed to be used entries, so no free entries should be there.

While most PDF viewers/libraries don't have an issue with free entries
in the initial revision, some do. Interestingly, those viewers/libraries
seem to work fine with the pre-1.0.1 way of creating multiple
subsections where free entries are not explicitly listed.

Since the PDF specification mandates the use of a single subsection for
the initial revision, the solution to avoid problems is to use the
optimize task with 'compact: true' automatically. This can be turned off
by setting the new argument 'compact' of Document#write to false which
should be carefully considered since it might lead to invalid documents.