Add Type::Annotations::Square and associated methods
The name "square" is a bit of a misnomer by the PDF specification as it
actually allows rectangle annotations.

Since square and circle (i.e. ellipse) annotations are very similar,
they share the same base class SquareCircle.

The needed method for creating a square annotation is appropriately
called #create_rectangle and follows the same semantics as
Canvas#rectangle.