Revert commit 093b1e5b to make text extraction in macOS Preview work
Text extraction works by mapping the character code to Unicode using the
/ToUnicode CMap. So the character code is taken, looked up in the CMap
and the result is one or more Unicode characters.

macOS Preview, however, has a bug and takes an invalid approach. It
first maps the character code to the respective CID via the /Encoding
entry and then uses this CID to look up the Unicode character(s) in the
/ToUnicode CMap.

Whenever the character code and the CID are not the same, invalid
results appear. Due to the performance optimization of not using certain
character codes (e.g. 13) to avoid escaping in the serialized literal
strings, the mapping of character codes to CIDs is not an identity
mapping.

So to get macOS Preview working correctly, we need to get to a situation
where character code = CID. And this is only possible by reverting the
commit which removed the special handling of certain subset glyph
IDs. Now this special handling is implemented when creating character
codes and subset glyph IDs and thus we have an identity mapping again.

Note that there is still one case where this doesn't work: When multiple
different Unicode characters map ot the same glyph (think hyphen and
soft-hyphen).

See https://github.com/gettalong/hexapdf/discussions/357 and
https://stackoverflow.com/questions/79560847/ for more details.