Add Layout::Style#font_bold and Layout::Style#font_italic
Currently setting a variant of a font is only possible by specifying the
variant together with the font using Style#font.

This means it is not easily possible to switch the font variant without
also knowing the name. And it is not possible to do this independently.

For example, if you want to "nest" styles by using one as the base style
and then merging the inner one, the inner style can't just change the
font variant to bold.

By introducing font_bold and font_italic style properties such scenarios
are now possible.