Fix AcroForm::Form#perform_validation to ensure only field type classes are used
If a field contains an embedded widget, it is automatically assigned the
widget annotation type class when loaded due to /Type and /Subtype.
However, in an AcroForm field hierarchy only field objects should appear.

This change ensures that all widget annotations that are also fields get
stored using a field type class in the field hierarchy.