Add Layout::ContainerBox#splitable
In the current implementation of the container box all child boxes have
to fit into one region. This is useful if all child boxes need to be
kept together.

However, sometimes it is useful to just wrap all child boxes into a
parent box so that the parent box itself can be styled (similar to how
<div> tags are used). And for this usage the container box needs to be
splitable.

To support both use-cases the new #splitable accessor is added.