Don't accidentally set subscript/superscript values in Layout::Style
The subscript and superscript values are used in various places: When
setting the values, when calculating the font size and when calculating
the text rise.

Currently ony the #subscript and #superscript methods are used. However,
by using them the values are accidentally set. What one really wants to
do is using `subscript? && subscript` (likewise with #superscript).