Fix encryption when using AES with 256bits and an owner password
For all standard security handler revisions R <= 4 (i.e. everything
except the combination of AES with 256bits) the value of the /O key
needs to be available when calculating the /U key.

However, for R=6, ie. AES with 256bits, it is the other way around: The
/U key needs to be available when calculating the /O key.

This was not taken into account of when implementing support for R=6 and
thus the owner key was not correctly computed and therefore unusable.