Change Dictionary validation to delete fields with invalid types
If a field entry has an invalid type (e.g. value of 20 for a field of
type String), HexaPDF currently just throws a validation error.

Since an invalid type is worse than having no field value at all in most
cases (because every code path would need to account for that), this
behaviour is changed to delete the value.